# switchboard/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Treadmill Switchboard API
  version: 0.1.0
paths:
...
  /jobs:
    get:
      tags:
      - Jobs
      summary: List jobs
      operationId: listJobs
      parameters:
      - in: query
        name: cursor
        description: Opaque keyset cursor from a previous response's `next_cursor`.
        schema:
          description: Opaque keyset cursor from a previous response's `next_cursor`.
          type: string
        style: form
      - in: query
        name: limit
        description: |-
          Maximum number of jobs per page. Omitted or out-of-range values fall
          back to the server's default and bounds.
        schema:
          description: |-
            Maximum number of jobs per page. Omitted or out-of-range values fall
            back to the server's default and bounds.
          type: integer
          format: uint32
          minimum: 0
        style: form
      responses:
        '401':
          description: 'Authentication failed: the bearer token is missing, malformed, expired, or revoked.'
        '403':
          description: The authenticated account is locked, or lacks permission for this resource.
        '200':
          description: |-
            Response body of `GET /jobs`: a page of jobs the caller can read, newest
            first.

            Pagination is **keyset** on `(queued_at, job_id)` descending: when
            `next_cursor` is non-null, pass it back as the `cursor` query parameter to
            fetch the next page; a null `next_cursor` means the last page. There is no
            total count.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListResponse'
      security:
      - token: []
    post:
      tags:
      - Jobs
      summary: Enqueue a job
      operationId: enqueueJob
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobRequest'
        required: true
      responses:
        '401':
          description: 'Authentication failed: the bearer token is missing, malformed, expired, or revoked.'
        '403':
          description: The authenticated account is locked, or lacks permission for this resource.
        '400':
          description: Failed to parse the request body as JSON
          content:
            text/plain:
              schema:
                type: string
        '415':
          description: 'Expected request with `Content-Type: application/json`'
          content:
            text/plain:
              schema:
                type: string
        '422':
          description: Failed to deserialize the JSON body into the target type
          content:
            text/plain:
              schema:
                type: string
        '201':
          description: The job was enqueued; the response carries its id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnqueueJobResponse'
      security:
      - token: []
  /jobs/{id}/events:
    get:
      tags:
      - Jobs
      summary: List a job's audit events
      operationId: listJobEvents
      parameters:
      - in: path
        name: id
        description: The resource's unique identifier.
        required: true
        schema:
          description: The resource's unique identifier.
          type: string
          format: uuid
        style: simple
      - in: query
        name: cursor
        description: Opaque keyset cursor from a previous response's `next_cursor`.
        schema:
          description: Opaque keyset cursor from a previous response's `next_cursor`.
          type: string
        style: form
      - in: query
        name: limit
        description: |-
          Maximum number of events per page. Omitted or out-of-range values fall
          back to the server's default and bounds.
        schema:
          description: |-
            Maximum number of events per page. Omitted or out-of-range values fall
            back to the server's default and bounds.
          type: integer
          format: uint32
          minimum: 0
        style: form
      responses:
        '401':
          description: 'Authentication failed: the bearer token is missing, malformed, expired, or revoked.'
        '403':
          description: The authenticated account is locked, or lacks permission for this resource.
        '200':
          description: |-
            A page of rendered audit events for one entity, newest first.

            Paginated by keyset: when `next_cursor` is non-null, pass it back as the
            `cursor` query parameter to fetch the next (older) page; a null `next_cursor`
            means the last page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditFeedResponse'
      security:
      - token: []
  /jobs/{id}/nats-log-token:
    post:
      tags:
      - Jobs
      summary: Create a NATS log-streaming token for a job
      operationId: createJobNatsLogToken
      parameters:
      - in: path
        name: id
        description: The resource's unique identifier.
        required: true
        schema:
          description: The resource's unique identifier.
          type: string
          format: uuid
        style: simple
      responses:
        '401':
          description: 'Authentication failed: the bearer token is missing, malformed, expired, or revoked.'
        '403':
          description: The authenticated account is locked, or lacks permission for this resource.
        '200':
          description: |-
            Connection credentials for tailing/replaying a job's console logs over NATS,
            returned by `POST /jobs/{id}/nats-log-token`.

            The token is a short-lived **bearer** user JWT scoped to this job only: it
            may *subscribe* to the job's log subjects (`subject`) and its own inboxes
            (`inbox_prefix`), and *publish* to the slice of the JetStream API needed to
            run an ordered consumer against the job's stream (`stream`) — enough to
            replay stored history and then follow live. The client connects with the
            token string alone (no nkey seed) to whichever endpoint suits its transport:
            `websocket_url` for browsers, `nats_url` for native TCP clients. The same
            token authorizes either. The token only needs to be valid at connect time —
            an established NATS connection is not dropped when the JWT expires — so a
            client re-requests credentials when it next reconnects, after roughly
            `expires_in_secs`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NatsLogStreamCredentials'
        '503':
          description: Log streaming is not enabled on this deployment.
      security:
      - token: []
  /jobs/{id}:
    get:
      tags:
      - Jobs
      summary: Get a job
      operationId: getJob
      parameters:
      - in: path
        name: id
        description: The resource's unique identifier.
        required: true
        schema:
          description: The resource's unique identifier.
          type: string
          format: uuid
        style: simple
      responses:
        '401':
          description: 'Authentication failed: the bearer token is missing, malformed, expired, or revoked.'
        '403':
          description: The authenticated account is locked, or lacks permission for this resource.
        '200':
          description: |-
            The full server-side view of a single job, returned by `GET /jobs/{id}`.

            Covers the job's identity, ownership, lifecycle state, the spec it was
            enqueued with, and — once it has run — its placement and terminal outcome.
            Secret parameters are redacted (see [`JobParameterView`]).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobInfo'
      security:
      - token: []
    delete:
      tags:
      - Jobs
      summary: Terminate a job
      operationId: terminateJob
      parameters:
      - in: path
        name: id
        description: The resource's unique identifier.
        required: true
        schema:
          description: The resource's unique identifier.
          type: string
          format: uuid
        style: simple
      responses:
        '401':
          description: 'Authentication failed: the bearer token is missing, malformed, expired, or revoked.'
        '403':
          description: The authenticated account is locked, or lacks permission for this resource.
        '202':
          description: Termination was initiated.
        '204':
          description: The job was already finalized; nothing to do.
      security:
      - token: []
...
components:
  securitySchemes:
    token:
      type: http
      scheme: bearer
      description: 'A Treadmill user API token, presented as `Authorization: Bearer <token>`.'
  schemas:
...
    AuditFeedResponse:
      description: |-
        A page of rendered audit events for one entity, newest first.

        Paginated by keyset: when `next_cursor` is non-null, pass it back as the
        `cursor` query parameter to fetch the next (older) page; a null `next_cursor`
        means the last page.
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/RenderedAuditRow'
        next_cursor:
          description: Opaque cursor for the next page, or null on the last page.
          type:
          - string
          - 'null'
      required:
      - events
...
    EnqueueJobResponse:
      description: 'Response body of `POST /jobs`: the id assigned to the freshly enqueued job.'
      type: object
      properties:
        job_id:
          type: string
          format: uuid
      required:
      - job_id
...
    JobListResponse:
      description: |-
        Response body of `GET /jobs`: a page of jobs the caller can read, newest
        first.

        Pagination is **keyset** on `(queued_at, job_id)` descending: when
        `next_cursor` is non-null, pass it back as the `cursor` query parameter to
        fetch the next page; a null `next_cursor` means the last page. There is no
        total count.
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/JobSummary'
        next_cursor:
          description: Opaque cursor for the next page, or null on the last page.
          type:
          - string
          - 'null'
      required:
      - jobs
...
    JobRequest:
      type: object
      properties:
        host_tag_requirements:
          description: |-
            Host eligibility: the set of tags the chosen host must carry (as a
            superset) for this job to be assigned to it. Tags are opaque strings
            (`key=value` pairs or bare flags, by convention only), matched by
            containment against the host's tags.
          type: array
          default: []
          items:
            type: string
        init_spec:
          description: What kind of job this is.
          allOf:
          - $ref: '#/components/schemas/JobInitSpec'
        override_timeout:
          type:
          - string
          - 'null'
        owner:
          description: |-
            The subject (user or group) to own the enqueued job. Must be the caller
            itself or a group the caller is a member of; absent, ownership defaults
            to the caller. Ownership decides who can later read, stop, and manage the
            job.
          type:
          - string
          - 'null'
          format: uuid
          default: null
        parameters:
          description: |-
            A hash map of parameters provided to this job execution. These
            parameters are provided to the puppet daemon.
          type: object
          additionalProperties:
            $ref: '#/components/schemas/JobParameter'
        restart_policy:
          $ref: '#/components/schemas/RestartPolicy'
        ssh_keys:
          description: |-
            The set of initial SSH keys to deploy onto the image.

            The image's configuration of the Treadmill puppet daemon determines
            how and whether these keys will be loaded.
          type: array
          items:
            type: string
        target_requirements:
          description: |-
            Target (DUT) eligibility: an ordered array of requested targets, each a
            set of tags an attached DUT must carry (as a superset). The scheduler
            assigns each entry to a distinct attached target (DUT) on the chosen
            host. Empty requests no DUTs. Target tags do not affect image selection.
          type: array
          default: []
          items:
            type: array
            items:
              type: string
      required:
      - init_spec
      - ssh_keys
      - restart_policy
      - parameters
...
    NatsLogStreamCredentials:
      description: |-
        Connection credentials for tailing/replaying a job's console logs over NATS,
        returned by `POST /jobs/{id}/nats-log-token`.

        The token is a short-lived **bearer** user JWT scoped to this job only: it
        may *subscribe* to the job's log subjects (`subject`) and its own inboxes
        (`inbox_prefix`), and *publish* to the slice of the JetStream API needed to
        run an ordered consumer against the job's stream (`stream`) — enough to
        replay stored history and then follow live. The client connects with the
        token string alone (no nkey seed) to whichever endpoint suits its transport:
        `websocket_url` for browsers, `nats_url` for native TCP clients. The same
        token authorizes either. The token only needs to be valid at connect time —
        an established NATS connection is not dropped when the JWT expires — so a
        client re-requests credentials when it next reconnects, after roughly
        `expires_in_secs`.
      type: object
      properties:
        expires_in_secs:
          description: |-
            Seconds until the token's `exp`; re-request credentials after this
            elapses (only needed to open a *new* connection).
          type: integer
          format: uint64
          minimum: 0
        inbox_prefix:
          description: |-
            Inbox prefix the client **must** configure on its connection
            (`_INBOX.logs-<job-id>`): the token's subscribe permission covers only
            inboxes under this prefix, not the account-default `_INBOX.>`.
          type: string
        jetstream_domain:
          description: |-
            The server's JetStream domain, when it is configured with one; the
            client must address the JetStream API through it. Usually absent.
          type:
          - string
          - 'null'
        nats_url:
          description: |-
            Plain-TCP NATS client URL (e.g. `nats://nats.example:4222`), for native
            clients that speak the binary protocol. Nullable: a deployment may
            choose to expose only the WebSocket endpoint publicly, in which case
            only `websocket_url` is returned. Browsers cannot use this — they must
            use `websocket_url`.
          type:
          - string
          - 'null'
        stream:
          description: 'JetStream stream holding this job''s logs: `logs-<job-id>`.'
          type: string
        subject:
          description: |-
            Subject wildcard covering all of this job's log channels:
            `logs.<job-id>.>`.
          type: string
        token:
          description: Bearer user JWT authorizing the scope described above.
          type: string
        websocket_url:
          description: |-
            NATS **WebSocket** URL (e.g. `wss://nats.example:443`), for browser
            clients, which cannot speak the plain TCP protocol. Absent when the
            deployment does not expose a WebSocket listener; a browser client cannot
            stream logs against such a deployment.
          type:
          - string
          - 'null'
      required:
      - subject
      - stream
      - inbox_prefix
      - token
      - expires_in_secs
...
    RestartPolicy:
      description: |-
        How many times a job may be **automatically restarted** after it is dropped
        by its host, supplied with the job at enqueue (`POST /jobs`).

        Each automatic restart enqueues a successor that inherits one fewer; the
        count a running job still has left is reported back as [`RestartPolicyState`]
        on [`JobInfo`]. A `max_restarts` of `0` disables automatic restarts.
      type: object
      properties:
        max_restarts:
          description: Maximum number of automatic restarts to grant this job.
          type: integer
          format: uint32
          minimum: 0
      required:
      - max_restarts
...